+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 2.2.6.0 (upcoming)
 
 * Upgrade to Apache Cassandra 2.2.6
+* Add transformation for getting the bounding box of a geographical shape
+
 
 ## 2.2.5.4 (24 May 2016)